Squashed commit of the following:

commit 4615dadb6a99cdff0b9d2b510fa11fd031ec2de8
Author: Steve Chew <stechew@cisco.com>
Date:   Tue Jul 13 14:47:21 2021 -0400

    Update

commit 9976e20f54ce0ca6d2d3d3906dd0bb1375bec726
Author: Steve Chew <stechew@cisco.com>
Date:   Fri Jun 25 16:38:13 2021 -0400

    events: Use instance_id to make event_id unique across threads.

+++ src/main/snort_config.h
@@ -331,7 +331,7 @@ public:
     //------------------------------------------------------
     // FIXIT-L command line only stuff, add to conf / module
 
-    uint32_t event_log_id = 0;
+    uint16_t event_log_id = 0;
     SfCidr obfuscation_net;
     std::string bpf_filter;
     std::string metadata_filter;
@@ -545,7 +545,7 @@ public:
     { return run_flags & RUN_FLAG__INLINE_TEST; }
 
     // event stuff
-    uint32_t get_event_log_id() const
+    uint16_t get_event_log_id() const
     { return event_log_id; }
 
     bool process_all_events() const
